refactor: split rust docker cleanup helpers
